Fixed various new warnings produced by gcc 4.2.1.  No bugs

+++ src/pl-comp.c
@@ -290,7 +290,7 @@ typedef struct
 } var_table, *VarTable;
 
 #undef struct_offsetp
-#define struct_offsetp(t, f) ((int)((t*)0)->f)
+#define struct_offsetp(t, f) ((size_t)((t*)0)->f)
 #define sizeofVarTable(isize) (struct_offsetp(var_table, entry) + sizeof(int)*(isize))
 
 #define mkCopiedVarTable(o) copyVarTable(alloca(sizeofVarTable(o->isize)), o)
@@ -897,7 +897,7 @@ Finish up the clause.
 		 clause.variables*sizeof(word) +
 		 sizeofClause(clause.code_size) +
 		 sizeof(*cref) +
-		 (int)argFrameP((LocalFrame)NULL, MAXARITY));
+		 (size_t)argFrameP((LocalFrame)NULL, MAXARITY));
 					/* Needed for new frame arguments */
 
     cref = (ClauseRef)p;
